Rename helpers to resource_helpers
