+++ api_app/api/routes/shared_service_templates.py
@@ -9,7 +9,7 @@ from models.schemas.resource_template import ResourceTemplateInResponse, Resourc
 from models.schemas.shared_service_template import SharedServiceTemplateInCreate, SharedServiceTemplateInResponse
 from resources import strings
 from services.authentication import get_current_admin_user, get_current_tre_user_or_tre_admin
-from .helpers import get_current_template_by_name
+from .resource_helpers import get_current_template_by_name
 
 
 shared_service_templates_core_router = APIRouter(dependencies=[Depends(get_current_tre_user_or_tre_admin)])
